requirements.txt
absl-py==2.4.0
accumulation_tree==0.6.4
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-cors==0.8.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.2.0
attrs==25.4.0
av==16.1.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorful==0.5.8
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.5
ctranslate2==4.7.1
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
entrypoints==0.4
executing==2.2.1
factor_analyzer==0.5.1
faiss-cpu==1.13.2
faster-whisper==1.2.1
fastjsonschema==2.21.2
filelock==3.25.0
fire==0.7.1
flatbuffers==25.12.19
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.2.0
gast==0.7.0
google-api-core==2.30.0
google-auth==2.49.1
google-pasta==0.2.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.4.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter_client==8.8.0
jupyter-console==6.6.3
jupyter_core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kaggle==2.0.0
kagglehub==1.0.0
kagglesdk==0.1.16
keras==3.13.2
kiwisolver==1.4.9
lark==1.3.1
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.46.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
medmnist==3.0.2
mistune==3.2.0
ml_dtypes==0.5.4
monai==1.5.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
namex==0.1.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.5.4
notebook_shim==0.2.4
numba==0.64.0
numpy==2.4.2
nvidia-cusparselt-cu12==0.7.1
onnxruntime==1.24.3
openai-whisper==20250625
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
opentelemetry-api==1.40.0
opentelemetry-exporter-prometheus==0.61b0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.1
pandocfilters==1.5.1
papermill==2.7.0
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.2
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-spy==0.4.1
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-discovery==1.1.3
python-json-logger==4.0.0
python-slugify==8.0.4
pyudorandom==1.0.0
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.54.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
Send2Trash==2.1.0
shellingham==1.5.4
simpleitk==2.5.3
six==1.17.0
smart_open==7.5.1
soupsieve==2.8.3
stack-data==0.6.3
sympy==1.14.0
tdigest==0.5.2.2
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tensorflow==2.20.0
termcolor==3.3.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2026.2.24
tiktoken==0.12.0
timm==1.0.25
tinycss2==1.4.0
tokenizers==0.22.2
torch==2.10.0
torchaudio==2.10.0
torchinfo==1.8.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
virtualenv==21.2.0
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.6
widgetsnbextension==4.0.15
wrapt==2.1.1
yarl==1.23.0
zipp==3.23.0
gradio>=4.0.0
